manejo de excepciones

+++ Customer_Management/src/main/java/Customer_Management/Service/ClientService.java
@@ -1,10 +1,12 @@
 package Customer_Management.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import Customer_Management.Error.LocalNotFoundException;
 import Customer_Management.Model.ClientModel;
 import Customer_Management.repository.IClientRepository;
 
@@ -16,14 +18,30 @@ public class ClientService implements IClientService{
     IClientRepository repository;
 
     @Override
-    public ClientModel add(ClientModel model) {
+    public ClientModel add(ClientModel model){
+
         return repository.save(model);
+        
     }
 
     @Override
-    public boolean delete(int id) {
+    public boolean delete(int id) throws LocalNotFoundException {
+        try {
+            repository.deleteById(id);
+            return true;
+        } catch (Exception e) {
+
+            throw new LocalNotFoundException("Error al eliminar el cliente");
+        }
+        
+       /*  //Busca en base de datos
+        Optional<ClientModel> Cliente = repository.findById(id);
+        //Validación
+        if (!Cliente.isPresent()) {
+            throw new LocalNotFoundException("Esta id no existe");
+        }
         repository.deleteById(id);
-        return true;
+        return true; */
     }
 
     @Override
@@ -32,7 +50,7 @@ public class ClientService implements IClientService{
     }
 
     @Override
-    public ClientModel findById(int id) {
+    public ClientModel findById(int id) throws LocalNotFoundException {
         return repository.findById(id).get();
     }
